experiment - extract mapping iterator
